+++ spec/shameless/model_spec.rb
@@ -130,16 +130,20 @@ describe Shameless::Model do
   end
 
   describe '#reload' do
-    it 'lazily reloads base cell state' do
-      _, model = build_store
+    it 'lazily reloads all cells' do
+      model = build_model_with_cell
       instance = model.put(hotel_id: 1, room_type: 'roh', check_in_date: Date.today.to_s, net_rate: 90)
+
+      instance.meta.update(net_rate: 95)
 
       second_instance = model.where(hotel_id: 1).first
       second_instance.update(net_rate: 100)
+      second_instance.meta.update(net_rate: 110)
 
       instance.reload
       expect(instance.ref_key).to eq(1)
       expect(instance[:net_rate]).to eq(100)
+      expect(instance.meta[:net_rate]).to eq(110)
     end
   end