+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Unreleased
+
+* Include shard in underlying index name
+
 ### 0.3.1 (2016-11-18)
 
 * Use more static names for underlying database indices